Further ProductList calculator functionality - added varaibles to ProductList entity, as well as a services package that will contain the calculator functionality
