+++ src/main/java/app/persistence/OrderMapper.java
@@ -1,6 +1,7 @@
 package app.persistence;
 
 import app.entities.Product;
+import app.entities.ProductList;
 import app.entities.User;
 import app.exceptions.DatabaseException;
 import io.javalin.Javalin;
@@ -17,31 +18,7 @@ public class OrderMapper {
 
 
 
-    public static List<Product> calculateProductList (int carportWidth, int carportLength, boolean shed, ConnectionPool connectionPool) {
-        List<Product> productList = null;
-        int carportSquareMeters = carportWidth * carportLength;
-        return productList;
-    }
 
-    private static int calculatePosts (int carportLength) {
-        // All carports starts with 4 posts
-        int posts = 4;
-        int remainingLength = (carportLength * 10) - 4300; // 1000mm front unsupported + 3300mm until additional posts needed
-        int additionalPosts;
-        if (remainingLength % 3400 == 0) {
-            additionalPosts = remainingLength / 3400;
-        } else {
-            additionalPosts = remainingLength / 3400 + 1;
-        }
-        posts += (additionalPosts * 2);
-
-        return posts;
-    }
-
-    private static int calculateRoof (int carportSquareMeters) {
-        int roofPlates = 0;
-        return roofPlates;
-    }
 
     public static int getWidthByID (int carportWidthID, ConnectionPool connectionPool) throws DatabaseException {
         String sql = "SELECT * FROM carport_widths WHERE carport_width_id = ?";
